+++ src/resources/map/tileset.h
@@ -77,15 +77,15 @@ class Tileset final : public ImageSet
         int calcMemoryLocal() const override final
         {
             int sz = ImageSet::calcMemoryLocal() +
-                sizeof(Tileset);
+                static_cast<int>(sizeof(Tileset));
             const std::map<std::string, std::string>::const_iterator it_end =
                 mProperties.end();
             std::map<std::string, std::string>::const_iterator it =
                 mProperties.begin();
             while (it != it_end)
             {
-                sz += (*it).first.capacity() +
-                    (*it).second.capacity();
+                sz += static_cast<int>((*it).first.capacity() +
+                    (*it).second.capacity());
                 ++ it;
             }
             return sz;
