+++ src/resources/subimage.cpp
@@ -225,8 +225,8 @@ void SubImage::decRef()
 
 int SubImage::calcMemoryLocal() const
 {
-    int sz = sizeof(SubImage) +
-        sizeof(std::map<float, SDL_Surface*>) +
+    int sz = static_cast<int>(sizeof(SubImage) +
+        sizeof(std::map<float, SDL_Surface*>)) +
         Resource::calcMemoryLocal();
     return sz;
 }
